+++ R/list-linters.R
@@ -6,10 +6,7 @@
 #'
 #' @examples
 #' list_linters(".")
-list_linters <- function(path) {
-  if (missing(path) && is_testing()) {
-    path <- "."
-  }
+list_linters <- function(path = ".") {
   out <- c(
     # "absolute_path", # TODO: really broken, too many false positives, e.g #42
     "any_duplicated",