+++ NEWS.md
@@ -10,6 +10,8 @@
 * When using external rules with the `with-<pkg>` syntax, if the YAML file
   contains several rules separated by "---", then `flir` would only use the 
   first one. This is now fixed (#95). 
+  
+* `list_linters()` now uses `path = "."` by default (#99).
 
 # flir 0.5.0